Issue #121: Refactors: reorganized some modules.
