+++ dds_registration/views/billing_stripe.py
@@ -6,7 +6,7 @@ from django.http import Http404, HttpRequest
 from django.shortcuts import redirect, render
 
 from ..models import Membership, Payment, Registration
-from ..money import convert_from_stripe_units, get_stripe_amount_for_currency
+from .helpers.stripe_amounts import convert_from_stripe_units, get_stripe_amount_for_currency
 from .helpers.stripe_payments import get_stripe_client_secret
